Refactor: 회원가입시 isAdult값을 boolean값이 아닌 'kids' | 'adult'처럼 string값으로 변경
