+++ components/webtoon/webtoonSimilar/WebtoonSimilarItem.tsx
@@ -2,7 +2,7 @@ import Image from 'next/image';
 import Link from 'next/link';
 import styled from 'styled-components';
 
-import { AdultCheckTypes } from '../../../types/adult-check-types';
+import { AdultCheckTypes } from '../../../types/common/adult-check-types';
 
 import {
   addRecentWebtoonHandler,
