+++ types/UI/webtoon-search-types.ts
@@ -2,5 +2,5 @@ import { DataTypes } from '../common/webtoon-types';
 
 export interface WebtoonSearchTypes {
   webtoon: DataTypes;
-  isAdult?: boolean | null;
+  isAdult?: string | null;
 }
