+++ types/common/adult-check-types.ts
@@ -3,6 +3,6 @@ import { DataTypes } from './webtoon-types';
 export interface AdultCheckTypes {
   webtoon: DataTypes;
   rank?: number;
-  isAdult?: boolean | null;
+  isAdult?: string | null;
   id?: string | null;
 }
