+++ types/common/webtoon-types.ts
@@ -14,7 +14,7 @@ export interface DataTypes {
 
 export interface WebtoonArrTypes {
   webtoon: DataTypes[];
-  isAdult?: boolean | null;
+  isAdult?: string | null;
   id?: string | null;
 }
